requirements.txt
asgiref==3.7.2
Django
sqlparse
tzdata
gunicorn
Pillow
requests
djangorestframework
django-cors-headers
django-filter
